+++ constants.py
@@ -71,3 +71,8 @@ BOX_TEXT_THICKNESS = 1
 BOX_TEXT_SCALE = 0.4
 
 MOUSE_COLOR = (0, 0, 255)
+
+IPC_PORT = 8002
+
+TIME_GOOD_START_MS = 27000
+TIME_INTERESTING_START_MS = 55000
